Temporarily remove lots of stuff
