+++ Cargo.toml
@@ -8,7 +8,7 @@ description = "Composable views and associated data structures for nih-plug UIs 
 resolver = "2"
 members = [
     "xtask",
-    "examples/visualizers", "examples/peak_graph", "examples/histogram", "examples/test_new"
+    "examples/visualizers"
 ]
 
 [lib]
