--- examples/histogram/src/main.rs
@@ -1,6 +1,0 @@
-use histogram::HistogramPlugin;
-use nih_plug::prelude::*;
-
-fn main() {
-    nih_export_standalone::<HistogramPlugin>();
-}
