--- examples/peak_graph/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "peak_graph"
-version = "0.1.0"
-edition = "2021"
-description = "A peak graph built using Cyma"
-
-[lib]
-crate-type = ["cdylib", "lib"]
-
-[dependencies]
-nih_plug = { git = "https://github.com/robbert-vdh/nih-plug.git", features = ["assert_process_allocs", "standalone"] }
-cyma = { path = "../../" }
-nih_plug_vizia = { git = "https://github.com/robbert-vdh/nih-plug.git" }
-
-[profile.release]
-lto = "thin"
-strip = "symbols"
-
-[profile.profiling]
-inherits = "release"
-debug = true
-strip = "none"
