--- examples/peak_graph/src/main.rs
@@ -1,6 +1,0 @@
-use nih_plug::prelude::*;
-use peak_graph::PeakGraphPlugin;
-
-fn main() {
-    nih_export_standalone::<PeakGraphPlugin>();
-}
