--- examples/test_new/README.md
@@ -1,5 +1,0 @@
-# test_new
-
-Test for new data structure.
-
-Most code has been copied from the peak buffer example.
