--- examples/test_new/src/main.rs
@@ -1,6 +1,0 @@
-use nih_plug::prelude::*;
-use test_new::PeakGraphPlugin;
-
-fn main() {
-    nih_export_standalone::<PeakGraphPlugin>();
-}
