--- src/controls/mod.rs
@@ -1,1 +1,0 @@
-//! Views which allow changing the state of your plug-in.
