+++ src/visualizers/mod.rs
@@ -2,20 +2,12 @@
 
 mod graph;
 mod grid;
-mod histogram;
-mod lissajous;
-mod meter;
-mod oscilloscope;
 mod spectrum_analyzer;
 mod unit_ruler;
 mod waveform;
 
 pub use graph::*;
 pub use grid::*;
-pub use histogram::*;
-pub use lissajous::*;
-pub use meter::*;
-pub use oscilloscope::*;
 pub use spectrum_analyzer::*;
 pub use unit_ruler::*;
 pub use waveform::*;
